requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
youtube-transcript-api          # keep whatever latest build you installed
openai==1.3.7
pydantic==2.5.0
httpx